chore: clean up and refactor strategy panel
