+++ src/components/strategies-panel/index.ts
@@ -1,0 +1,4 @@
+export { StrategiesPanel } from './StrategiesPanel'
+export { StrategyTable } from './StrategyTable'
+export { StrategyRow } from './StrategyRow'
+export { StrategyAllocationChart } from './StrategyAllocationChart'
